Sprite texture batch - Events with unique ptrs

+++ Wolf/Wolf/assets/shaders/default_sprite/renderer2d.vert
@@ -2,7 +2,7 @@
 layout (location = 0) in vec3 a_pos;
 layout (location = 1) in vec4 a_color;
 layout (location = 2) in vec2 a_uv;
-layout (location = 3) in int a_texture_index;
+layout (location = 3) in float a_texture_index;
 
 uniform mat4 u_projection;
 uniform mat4 u_view;
@@ -11,7 +11,7 @@ out VS_OUT
 {
    vec4 color;
    vec2 uv;
-   flat int texture_index;
+   flat float texture_index;
 } vs_out;
