+++ Wolf/Wolf/rendering/batch/SpriteBatch.h
@@ -17,20 +17,20 @@ namespace Wolf
             glm::vec3 position;
             glm::vec4 color;
             glm::vec2 uv;
-            uint32_t texture_slot;
+            float texture_slot;
         };
 
         class SpriteBatch : public Batch
         {
             public:
-            void init(Shared<Material>& material) override;    
-			void shutdown() override;
-			void new_frame() override;
-			void end_frame() override;
-            void submit_primitive(const glm::vec3& position, const glm::vec3& size, const glm::vec4& color);
-            void submit_primitive(const glm::vec3& position, const glm::vec3& size, const glm::vec4& color, uint32_t texture_slot);
+                void init(Shared<Material>& material) override;    
+                void shutdown() override;
+                void new_frame() override;
+                void end_frame() override;
+                void submit_primitive(const glm::vec3& position, const glm::vec3& size, const glm::vec4& color);
+                void submit_primitive(const glm::vec3& position, const glm::vec3& size, const glm::vec4& color, uint32_t texture_slot);
             private:
-            void _flush();
+                void _flush();
 
             private:
             SpriteVertex* _buffer;
